+++ api/.env.example
@@ -1,9 +1,10 @@
 ##FOLDER PATH
 FOLDER_PATH= ''
 
-#AWS KEYS
+#AWS
 ACCESS_KEY = ''
 SECRET_KEY = ''
+BUCKET_NAME= ''
 
 ##MS TEAMS
 MS_WEBHOOK=''
