--- api/.gitignore
@@ -1,3 +1,0 @@
-json_data/*
-!json_data/teams.json
-.env
